separate enemy types from enemy data

+++ src/render_core/i_render_context.h
@@ -10,8 +10,8 @@ namespace cgt::render
 
 struct RenderStats
 {
-    usize spriteCount;
-    usize drawcallCount;
+    u32 spriteCount;
+    u32 drawcallCount;
 };
 
 class IRenderContext
